Don't ignore far agenda deadlines and skip all done timestamps
This commit makes the global TODO list not ignore deadlines,
(since the default setting for org-agenda-todo-ignore-deadlines is
nil) and makes the agenda ignore all done timestamps for a cleaner
view. The reason for removing ignore-deadlines is:
1) In the description of the variable, it says that the 'far
option is for when you want to use the global TODO list to
figure out what to do next. This is not exactly the case for me,
since I use NEXT keywords.
2) Right now I'm using the global DONE keyword to archive the entries,
and if it skips entries that had a deadline I have to search the files
manually which is tedious.
3) I rarely give TODO keywords deadlines (I use that mostly for
projects) so a couple more stuff in the global TODO list won't bother me.